Backend/requirements.txt
flask
flask-cors
scikit-learn
pandas
numpy
nltk
joblib
newspaper3k 
python-docx